requirements.txt
flask
flask-cors
iopaint
Pillow
torch
torchvision
opencv-python
numpy
